# postal_parse_address() performance