--- a/mismo/lib/geo/tests/.gitignore
+++ b/mismo/lib/geo/tests/.gitignore
@@ -1,0 +1,1 @@
+*.parquet