Not ready for merge, just opening to share the code with @NickCrews and @jstammers following discussion in #47 (is there a better way?).

To run the benchmarks just execute `python -m benchmarks.parse_address`.
This will print the timings on the screen and store the DuckDB files inside `benchmarks/db/<TIMESTAMP>`.

Here are my results:

```bash
$ python -m benchmarks.parse_address
_parse_fn__noop           took  14.6711 seconds
_parse_fn__python_only    took  15.4183 seconds
_parse_fn__postal_only    took  87.2642 seconds
_parse_fn__complete       took  88.1358 seconds
postal_parse_address      took 107.9780 seconds
```

My laptop specs:

```bash
OS: Ubuntu 21.04 x86_64 
Host: Latitude 5500 
Kernel: 5.11.0-49-generic 
CPU: Intel i7-8665U (8) @ 4.800GHz 
GPU: Intel WhiskeyLake-U GT2 [UHD Graphics 620] 
Memory: 8067MiB / 15813MiB 
```

_Note_: `_parse_fn__complete` has (at least) one bug.
When the `result` dictionary is converted to an ibis/DuckDB data structure, keys and values are shuffled (maybe the code that makes the conversion relies on the dictionary entries to have the same order as the fields declared in `ADDRESS_SCHEMA`?).
Here are the first rows of the table I get:

```bash
┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
┃ _parse_fn__complete_0(full_address)                                              ┃
┡━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┩
│ struct<street1: string, street2: string, city: string, state: string, postal_co… │
├──────────────────────────────────────────────────────────────────────────────────┤
│ {'street1': 'anchorage', 'street2': 'alaska', ... +4}                            │
│ {'street1': 'juneau', 'street2': 'alaska', ... +4}                               │
│ {'street1': 'anchorage', 'street2': 'alaska', ... +4}                            │
│ {'street1': 'anchorage', 'street2': 'alaska', ... +4}                            │
│ {'street1': 'anchorage', 'street2': 'alaska', ... +4}                            │
└──────────────────────────────────────────────────────────────────────────────────┘
```