@jstammers 2024-07-15
This looks very promising, thanks for starting this @lmores .

I wonder if we can avoid the problem you've highlighted by explicitly returning them in order at the end

```python
{
    "street1": result["street1"],
    "street2": result["street2"],
    ...
}
```
although this feels unnecessary as I would expect the fields to be inferred correctly from the schema. 
This could be an upstream bug with ibis - here's what I see with a toy example where I reverse the order of the fields in the address

```python
import ibis
from ibis.expr import datatypes as dt
ibis.options.interactive = True

import pandas as pd

ADDRESS_SCHEMA = dt.Struct(
    {
        "street1": "string",
        "street2": "string",
        "city": "string",
        "state": "string",
        "postal_code": "string",
        "country": "string",
    }
)

@ibis.udf.scalar.python
def reversed_address() -> ADDRESS_SCHEMA:
    return {
        "country": "country",
        "postal_code": "postal_code",
        "state": "state",
        "city": "city",
        "street2": "street2",
        "street1": "street1",
    }

@ibis.udf.scalar.python
def address() -> ADDRESS_SCHEMA:
    return {
        "street1": "street1",
        "street2": "street2",
        "city": "city",
        "state": "state",
        "postal_code": "postal_code",
        "country": "country",
    }

table = ibis.memtable(pd.DataFrame({"ix":[1,2,3,4]}))
table.mutate(address=address(), rev_address=reversed_address())

┏━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
┃ ix    ┃ address                                                                          ┃ rev_address                                                                      ┃
┡━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┩
│ int64 │ struct<street1: string, street2: string, city: string, state: string, postal_co… │ struct<street1: string, street2: string, city: string, state: string, postal_co… │
├───────┼──────────────────────────────────────────────────────────────────────────────────┼──────────────────────────────────────────────────────────────────────────────────┤
│     1 │ {'street1': 'street1', 'street2': 'street2', ... +4}                             │ {'street1': 'country', 'street2': 'postal_code', ... +4}                         │
│     2 │ {'street1': 'street1', 'street2': 'street2', ... +4}                             │ {'street1': 'country', 'street2': 'postal_code', ... +4}                         │
│     3 │ {'street1': 'street1', 'street2': 'street2', ... +4}                             │ {'street1': 'country', 'street2': 'postal_code', ... +4}                         │
│     4 │ {'street1': 'street1', 'street2': 'street2', ... +4}                             │ {'street1': 'country', 'street2': 'postal_code', ... +4}                         │
└───────┴──────────────────────────────────────────────────────────────────────────────────┴──────────────────────────────────────────────────────────────────────────────────┘
```

Maybe for now, the solution is to explicitly enforce the order of the fields in the returned object.


Another edge-case to test would be one in which multiple values for the same field are returned. Here's an example for an address that looks correct, but because the `city` and `county` are swapped, the parsed result contains multiple `city` components

```python
parse_address("123 Elm Street, Los Angeles, Burbank, CA 91501")
[('123', 'house_number'),
 ('elm street', 'road'),
 ('los', 'city'),
 ('angeles', 'suburb'),
 ('burbank', 'city'),
 ('ca', 'state'),
 ('91501', 'postcode')]
```
I expect these are more likely to be invalid parses than correct ones, but worth noting.