@NickCrews 2024-07-17
Thanks for this @lmores , this looks awesome!

That is a bug with ibis, I filed an issue at https://github.com/ibis-project/ibis/issues/9613. If possible, can we keep this PR/discussion just related to perf/benchmarking as much as possible?

Here is my interpretation of the benchmarks:

- _parse_fn__python_only - _parse_fn__noop ~= .9s
- _parse_fn__complete - _parse_fn__postal_only ~= .9s
- _parse_fn__postal_only - _parse_fn__noop ~= 73s
- _parse_fn__complete - _parse_fn__python_only ~= 73

so what this says to me is the time of this UDF can be accounted as the sum of
- baseline overhead of a UDF (~14.6s)
- our python logic (.9s)
- pypostal running (73s)

What do you think of this interpretation?

If this is correct, then trying to optimize the python is going to be fruitless, and we should just go with the most idiomatic version we can.

Even if we don't change our implementation, I still think this benchmarking harness code is useful and we should add it