@lmores 2024-07-17
@NickCrews: I agree with your interpretation of the results, but not on the conclusions XD; and here is why: all functions named `_parse_fn__*` have been obtained by modifying my implementation of the UDF, not the current one! (my bad, should have mentioned it).

Timings must be read as follows:

```text
$ python -m benchmarks.parse_address
_parse_fn__noop           took  14.6711 seconds
_parse_fn__python_only    took  15.4183 seconds    # my implementation, not calling pypostal
_parse_fn__postal_only    took  87.2642 seconds
_parse_fn__complete       took  88.1358 seconds    # my implementation
postal_parse_address      took 107.9780 seconds    # current implementation
```

Since current implementation (`postal_parse_address`) takes `107` seconds and my implementation (`_parse_fn__complete`) takes `88` seconds, we would gain  `~20` seconds switching to the new one (and your analysis proves that we cannot improve anymore with plain python as `_parse_fn__complete - _parse_fn__postal_only ~= .9s`).

Does it make sense?