@NickCrews 2024-07-17
FYI I just pushed https://github.com/NickCrews/mismo/commit/ac10b6c60f0e8069f882f3cdfbe0d3f59023c784 that adjusts the current implementation a bit, but shouldn't affect runtime

I am looking back at my old notes from https://github.com/NickCrews/mismo/issues/19, and I think I want to take this opportunity to actually use pytest-benchmark for this. So I'm tweaking this PR right now locally to use this framework, so hold off on any edits you're making. Thanks!