@NickCrews 2024-07-17
lol, actually we already have pytest-benchark set up. grep for `benchmark` and you can find examples. Still, I'm migrating this PR to use that.