@NickCrews 2024-07-17
nevermind I was able to remove the WIP.

If you want, I'd love to see you actually change the implementation to the faster version in a followup PR!